+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Reduced the cost of Colonels for the CEO Commander
 - Atlas upgrade increases Atlas health by 100% instead of 25%
 - Locust upgrade now adds splash damage rather than vision
+- Lob Upgrade doubles the fire rate and removes the recharge cost
 
 ## v5.61.0 - 2024-03-13