+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Atlas Upgrade Tech health buff increased
 - Locust Upgrade Tech now adds splash damage rather than vision
 - Lob Upgrade Tech doubles the fire rate and removes the recharge cost
+- Omega Upgrade Tech now also doubles the rate of fire
 
 ## v5.61.0 - 2024-03-13